+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-## Next Version (Unreleased)
+## Next version (Unreleased)
+
+FEATURES:
+
+IMPROVEMENTS:
+
+- commands/login: Add support for two-factor authentication [GH-8935]
+- commands/ssh-config: Properly display windows path if invoked from msys2 or cygwin [GH-8915]
+- providers/salt: Remove duplicate stdout, stderr output from salt [GH-8767]
+- providers/salt: Introduce salt_call_args and salt_args option for salt provisioner [GH-8927]
+
+BUG FIXES:
+
+- guests/shell_expand_guest_path : Properly expand guest paths that include relative path alias [GH-8918]
+- util/ssh: Properly quote key path for IdentityFile option to allow for spaces [GH-8924]
+
+## 1.9.8 (August 23, 2017)
 
 IMPROVEMENTS: